Add check_infinite_loop to its own module
